# framework/misc/orchestratorAPISpec.yml
openapi: '3.0.0'
info:
  description: "This is the specification for the API of the intueri-orchestrator system."
  version: "1.0.0-SNAPSHOT"
  title: "intueri-orchestrator"
  contact:
    email: "speckij@posteo.net"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
  /detectors:
    get:
      tags:
      - "detector"
      summary: "Gets a list of all detectors"
      description: "Returns an Array of all detectors with their last known status"
      operationId: "getDetectors"
      responses:
        200:
          description: "successfully retrieved detectors"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/DetectorSchema'
        500:
          description: "error while retrieving detectors"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSchema'
...
  /rules:
    get:
      tags:
      - "rules"
      summary: "Gets a list of all Rules"
      description: "Returns an Array of all Rules"
      operationId: "getRules"
      responses:
        200:
          description: "successfully retrieved rules"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/RuleSchema'
        500:
          description: "error while retrieving rules"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSchema'
    post:
      tags: 
      - "rules"
      summary: Creates a new Rule
      description: "Creates a new Rule. The id field should be empty to allow for autogeneration of the id by the server. If an id is specified an a configuration with the id already exists the old 
        configuration is overwritten."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleSchema'
      responses:
        200:
          description: "successfully created rule"
          
        400:
          description: "Error occured while creating rule. Request was invalid."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSchema'
        500:
          description: "Error occured while creating rule. Server error."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSchema'
  /rules/{id}:
    put:
      tags: 
      - "rules"
      summary: Updates an existing Rule.
      description: "Updates an existing Rule. The Schema of the Rule is autogenerated based on the schema of the datastore."
      parameters:
      - name: id
        in: path
        description: id of the rule to be updated
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleSchemaWithId'
            example:
            - id: DemoRule
              version: 1
              relation: value_blob
              output:
               - type: kafka
                 options:
                   topic: outputtopic
              filters:
               - type: COMPARISON
                 column: value_quantity
                 options:
                   operator: EQ
                   value: '23'
                   
      responses:
        200:
          description: "successfully updated configuration"
          
        400:
          description: "Error occured while updating configuration. Request was invalid."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSchema'
        500:
          description: "Error occured while updating configuration. Server error."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSchema'
    delete:
      tags: 
      - "rules"
      summary: Deletes a Rule
      description: Deletes a Rule.
      parameters:
      - name: id
        in: path
        description: id of the rule to be deleted
        required: true
        schema:
          type: string
      responses:
        200:
          description: "successfully deleted rule"
        400:
          description: "Error occured while deleting rule. Request was invalid."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSchema'
        500:
          description: "Error occured while deleting rule. Server error."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorSchema'
#    post:
#      tags:
#      - "pet"
#      summary: "Add a new pet to the store"
#      description: ""
#      operationId: "addPet"
#      consumes:
#      - "application/json"
#      - "application/xml"
#      produces:
#      - "application/xml"
#      - "application/json"
#      parameters:
#      - in: "body"
#        name: "body"
#        description: "Pet object that needs to be added to the store"
#        required: true
#        schema:
#          $ref: "#/definitions/Pet"
#      responses:
#        405:
#          description: "Invalid input"
#      security:
#      - petstore_auth:
#        - "write:pets"
#        - "read:pets"
#    put:
#      tags:
#      - "pet"
#      summary: "Update an existing pet"
#      description: ""
#      operationId: "updatePet"
#      consumes:
#      - "application/json"
#      - "application/xml"
#      produces:
#      - "application/xml"
#      - "application/json"
#      parameters:
#      - in: "body"
#        name: "body"
#        description: "Pet object that needs to be added to the store"
#        required: true
#        schema:
#          $ref: "#/definitions/Pet"
#      responses:
#        400:
#          description: "Invalid ID supplied"
#        404:
#          description: "Pet not found"
#        405:
#          description: "Validation exception"
#      security:
#      - petstore_auth:
#        - "write:pets"
#        - "read:pets"
components:
  schemas:
    ErrorSchema:
      type: "object"
      additionalProperties: false
      properties:
        errorMessage:
          type: "string"
          example: "Oh noes an error has occurred!"
...
    DetectorSchema:
      type: object
      required:
      - id
      - name
      - status
      - schema
      - config
      - enabledRules
      - availableRules
      properties:
        id:
          type: string
          title: The unique Id of the detector
          example: "c08364df-628e-479e-8b7f-3614cc9a8aca"
          pattern: "^(([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12})$"
        name:
          type: string
          title: The human readable name of the detector
          example: "Detector on Primary Datastore XY123"
        status:
          type: string
          title: The current status of the detector
          enum:
          - "WAITING_FOR_CONFIG"
          - "UPDATING_SCHEMA"
          - "WAITING FOR RULES"
          - "BOOTSTRAPPING"
          - "PAUSED"
          - "RUNNING"
          - "ERROR"
        schemaId:
          type: string
          title: The unique Id of the schema used in the datastore of this detector
          example: "d8fb57f1-f922-44a1-a5de-20465a5f3ce0"
          pattern: "^(([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12})$"
        config:
          type: string
          title: The unique Id of the config of this detector
          pattern: "^(([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12})$"
          example: "9c90817d-a107-4a78-ab71-1057301e2fb9"
        enabledRules:
          type: array
          items:
            type: string
          title: The unique Ids of all rules currently applied by this detector
          pattern: "^(([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12})$"
          example: 
          - "0b7916d7-99e6-47d7-aaa6-b4387daa8781"
          - "e59e7e8e-9e71-499b-b88f-92bb3b8c8c0f"
        availableRules:
          type: array
          items:
            type: string
          title: The unique Ids of all rules currently applied by this detector
          pattern: "^(([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12})$"
          example: 
          - "0b7916d7-99e6-47d7-aaa6-b4387daa8781"
          - "e59e7e8e-9e71-499b-b88f-92bb3b8c8c0f"
...
    RuleSchema:
      type: object
      description: Schemas that all rules applicable to this detector have to conform to. Is autogenerated based on the schema of the datastore.
    RuleSchemaWithId:
      allOf:
      - $ref: '#/components/schemas/RuleSchema'
      - type: object
        required:
        - id